Implement an iterator for postorder traversals.
I believe that this will be useful in pre-processing passes, many of
which do postorder traversals that they implement by hand.

Right now this iterator does not support modification of the traversal
pattern, but we could add this later on, if we want it.

This iterator aims to be STL compliant, but I'm not an expert.