Always mark propagations in incremental mode
Fixes #3991. In incremental mode, the logic can change from one
`(check-sat)` call to another. In the reported issue, we start with
`QF_NIA` but then switch to `QF_UFNIA` because there is a `div` term
(which has a UF in its expanded form). Thus, we have to be careful when
we check whether sharing is enabled or not. In the issue, we were not
marking propagations because sharing was not enabled at the time but
then were expecting to have that information available later on. This
commit fixes the issue by always marking propagations in incremental
mode. Other solutions are possible: we could widen the theory eagerly in
incremental mode or we could try to make `getExplanationAndRecipe()`
robust to this case.