Collect statistics about normal form inferences
This commit adds code to count the number of inferences made of each
inference type for normal form inferences. It extends the `Inference`
enum in `infer_info.h` and adds two new `sendInference()` methods in the
`InferenceManager` to send and count inferences that have a corresonding
entry in the `Inference` enum.