Support async-signal-safe printing of inferences
Commit c9b7c3d6fcd49b6d75a85e1316e9918374d1ebbe introduced code for
counting the number of string inferences done per type of inference.
However, it did not add support for printing the inference names in an
async-signal-safe manner via `safe_print()` (note that printing in
signal handlers is done differently from the regular stats printing).
This commit adds the corresponding code, s.t. we get the inference names
when printing the stats when CVC4 is interrupted or crashes.