All commented out code and changes needed to remove -Wshadow warning
Originally thought it was only expand_list_arg and wrote the message below (I was wrong)

Semantics-preserving change to python bindings expand_list_arg

The generated Cython code causes a -Wshadow warning before this change
I didn't dig into it much, but it appears to be related to a generated
case statement when there's computation on arguments from a closure
(e.g. num_req_args in that function). The case statement has unreachable
code and shadows some variables. This minor rewrite does the computation
on the local variable