Add safe_print() support for Kind enum
This commit changes the `mkkind` script to generate a `toString()`
method for the `Kind` enum. This method can be used by the
`safe_print()` function to print statistics if CVC4 has been terminated
before solving a problem. The stats for strings include statistics that
rely on printing kinds (e.g. the number of reductions done of each
kind).